+++ crawl-ref/source/religion.cc
@@ -322,7 +322,7 @@ const vector<god_power> god_powers[NUM_GODS] =
            "The storm surrounding you is now too weak to repel missiles." },
       { 4, "You adapt resistances upon receiving elemental damage.",
            "You no longer adapt resistances upon receiving elemental damage." },
-      { 5, ABIL_QAZLAL_DISASTER_AREA, "release nature's wrath around you" },
+      { 5, ABIL_QAZLAL_BECOME_THE_STORM, "become lightning and release nature's wrath" },
     },
 
     // Ru
@@ -2753,7 +2753,8 @@ void excommunication(bool voluntary, god_type new_god)
         you.stop_train.insert(abil_skill(abil));
 
     update_can_train();
-    you.can_train.set(SK_INVOCATIONS, false);
+    if (you.species != SP_HIGH_ELF)
+        you.can_train.set(SK_INVOCATIONS, false);
     reset_training();
 
     // Perhaps we abandoned Trog with everything but Spellcasting maxed out.