+++ crawl-ref/source/spl-cast.cc
@@ -2009,6 +2009,9 @@ static spret _do_cast(spell_type spell, int powc, const dist& spd,
     case SPELL_POISONOUS_VAPOURS:
         return cast_poisonous_vapours(powc, spd, fail);
 
+    case SPELL_BLINKBOLT:
+        return blinkbolt(powc, beam, fail);
+
     // non-player spells that have a zap, but that shouldn't be called (e.g
     // because they will crash as a player zap).
     case SPELL_DRAIN_LIFE: